Moving queries to common/queries
