+++ components/shopify_partner/common/queries/getAppUninstalls.mjs
