+++ components/shopify_partner/sources/new-app-uninstalls/new-app-uninstalls.mjs
@@ -1,6 +1,6 @@
 import common from "../../common/common.mjs";
 import shopify from "../../shopify_partner.app.mjs";
-import getAppUninstalls from "../../queries/getAppUninstalls.mjs";
+import getAppUninstalls from "../../common/queries/getAppUninstalls.mjs";
 
 export default {
   key: "shopify_partner-new-app-uninstalls",
